% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{failure}
\alias{failure}
\title{Wraps a value in \code{failure} type of \code{result}}
\usage{
failure(value = "failed", status = "error")
}
\arguments{
\item{value}{any object to wrap}

\item{status}{character string of the result (typically short)}
}
\value{
\code{result} object of subclass \code{failure}
}
\description{
Wraps a value in \code{failure} type of \code{result}
}
\examples{
failure()
failure(42)
}
