% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_failure}
\alias{is_failure}
\title{Checks if an object is of \code{failure} class}
\usage{
is_failure(obj)
}
\arguments{
\item{obj}{object to check}
}
\value{
\code{TRUE} if object is of \code{failure} class,
\code{FALSE} otherwise
}
\description{
Checks if an object is of \code{failure} class
}
\examples{
is_failure(success())
is_failure(failure())
}
