% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_result}
\alias{is_result}
\title{Checks if an object is of \code{result} class}
\usage{
is_result(obj)
}
\arguments{
\item{obj}{object to check}
}
\value{
\code{TRUE} if object is of \code{result} class,
\code{FALSE} otherwise
}
\description{
Checks if an object is of \code{result} class
}
\examples{
is_result(success())
is_result(failure())
is_result(42)
}
