% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_result_monad}
\alias{is_result_monad}
\title{Checks if an object is of \code{result.monad} class}
\usage{
is_result_monad(obj)
}
\arguments{
\item{obj}{object to check}
}
\value{
\code{TRUE} if object is of \code{result.monad} class,
\code{FALSE} otherwise
}
\description{
Checks if an object is of \code{result.monad} class
}
\examples{
is_result_monad(result(mean))  # TRUE
is_result_monad(success())     # FALSE
is_result_monad(42)            # FALSE
}
