% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_try.R
\name{or_try}
\alias{or_try}
\title{Tries another result if previous one failed (meant to be used with pipe)}
\usage{
or_try(last_result, next_obj, ...)
}
\arguments{
\item{last_result}{\code{result} object of subclass \code{success} or
\code{failure}}

\item{next_obj}{\code{result} monad or plain function to bind with}

\item{...}{additional arguments to pass to \code{next_obj}}
}
\value{
\code{result} object of subclass \code{success} or \code{failure}
}
\description{
Tries another result if previous one failed (meant to be used with pipe)
}
\examples{
bad <- function() stop("bad")
good <- function() "good"
also_good <- function() "also good"
also_bad <- function() stop("also bad")

res <- as_result(bad()) |> or_try(good) |> or_try(also_good)
is_success(res)    # TRUE
value(res)         # "good"
status(res)        # "ok"

safely_bad <- result(bad)
res <- safely_bad() |> or_try(good) |> or_try(also_good)
is_success(res)    # TRUE
value(res)         # "good"
status(res)        # "ok"
}
