% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{success}
\alias{success}
\title{Wraps a value in \code{success} type of \code{result}}
\usage{
success(value = "done", status = "ok")
}
\arguments{
\item{value}{any object to wrap}

\item{status}{character string of the result (typically short)}
}
\value{
\code{result} object of subclass \code{success}
}
\description{
\code{success} is a constructor function for \code{result} class.
}
\examples{
success()
success(42)
}
