% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{value}
\alias{value}
\title{Extracts the value of a \code{result}}
\usage{
value(obj)
}
\arguments{
\item{obj}{\code{result} object}
}
\value{
value object wrapped by \code{result}
}
\description{
Extracts the value of a \code{result}
}
\examples{
value(success(42))
}
